requirements.txt
streamlit==1.31.1
opencv-python-headless==4.9.0.80
numpy==1.26.4
Pillow==10.2.0
pyyaml==6.0.1
python-dotenv==1.0.0

